--- toolkit-api/test/services/health/health.test.ts
@@ -1,11 +1,0 @@
-// For more information about this file see https://dove.feathersjs.com/guides/cli/service.test.html
-import assert from 'assert'
-import { app } from '../../../src/app'
-
-describe('health service', () => {
-  it('registered the service', () => {
-    const service = app.service('health')
-
-    assert.ok(service, 'Registered the service')
-  })
-})
